add hero banner with animations
