+++ package.json
@@ -21,6 +21,7 @@
     "clsx": "^2.0.0",
     "eslint": "8.46.0",
     "eslint-config-next": "13.4.13",
+    "framer-motion": "^10.16.4",
     "lucide-react": "^0.265.0",
     "next": "13.4.13",
     "openai-edge": "^1.2.2",
