+++ src/app/layout.tsx
@@ -14,7 +14,7 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body className={cn(inter.className, 'bg-gray-950')}>
+      <body className={cn(inter.className, 'bg-[#010A2B]')}>
         <header className="flex justify-center items-center h-20 text-white">
           <KvikkLogo className="h-4 w-4 fill-white" />
         </header>
