+++ src/assets/Icons.tsx
@@ -1,64 +1,118 @@
-import { cn } from '@/lib/utils';
+import { cn } from "@/lib/utils";
 
 type Props = {
-  className: string;
+    className: string;
 };
 
 export function Check({ className }: Props) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={cn(className)}
-    >
-      <polyline points="20 6 9 17 4 12" />
-    </svg>
-  );
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className={cn(className)}
+        >
+            <polyline points="20 6 9 17 4 12" />
+        </svg>
+    );
 }
 export function ChevronDown({ className }: Props) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={cn(className)}
-    >
-      <path d="m6 9 6 6 6-6" />
-    </svg>
-  );
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className={cn(className)}
+        >
+            <path d="m6 9 6 6 6-6" />
+        </svg>
+    );
 }
 export function Clipboard({ className }: Props) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={cn(className)}
-    >
-      <rect width="8" height="4" x="8" y="2" rx="1" ry="1" />
-      <path d="M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" />
-      <path d="M16 4h2a2 2 0 0 1 2 2v4" />
-      <path d="M21 14H11" />
-      <path d="m15 10-4 4 4 4" />
-    </svg>
-  );
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className={cn(className)}
+        >
+            <rect width="8" height="4" x="8" y="2" rx="1" ry="1" />
+            <path d="M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" />
+            <path d="M16 4h2a2 2 0 0 1 2 2v4" />
+            <path d="M21 14H11" />
+            <path d="m15 10-4 4 4 4" />
+        </svg>
+    );
+}
+export function Mail({ className }: Props) {
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className={cn(className)}
+        >
+            <rect width="8" height="4" x="8" y="2" rx="1" ry="1" />
+            <path d="M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" />
+            <path d="M16 4h2a2 2 0 0 1 2 2v4" />
+            <path d="M21 14H11" />
+            <path d="m15 10-4 4 4 4" />
+        </svg>
+    );
+}
+export function AiMail({ className }: Props) {
+    return (
+        <svg
+            width="58"
+            height="55"
+            viewBox="0 0 58 55"
+            xmlns="http://www.w3.org/2000/svg"
+            className={cn("fill-white h-10 w-10", className)}
+        >
+            <path d="M56.6578 26.4475H31.2544C30.9659 26.4475 21.2858 26.5599 21.0818 26.76C20.8778 26.96 20.7632 27.2314 20.7632 27.5143V51.6951C20.7632 52.3553 21.0306 52.9884 21.5065 53.4552C21.9825 53.922 22.6281 54.1843 23.3012 54.1843H55.2076C55.8807 54.1843 56.5262 53.922 57.0022 53.4552C57.4782 52.9884 57.7456 52.3553 57.7456 51.6951V27.5143C57.7456 27.2314 57.631 26.96 57.427 26.76C57.223 26.5599 56.9463 26.4475 56.6578 26.4475ZM39.2544 41.7134L24.6463 28.5811H53.8624L39.2544 41.7134ZM34.4811 40.3159L22.9386 50.6923V29.9395L34.4811 40.3159ZM36.0909 41.7632L38.5292 43.9466C38.7297 44.1265 38.9917 44.2263 39.2634 44.2263C39.5352 44.2263 39.7971 44.1265 39.9976 43.9466L42.4269 41.7632L53.8624 52.0507H24.6481L36.0909 41.7632ZM44.0276 40.3159L55.5701 29.9395V50.6923L44.0276 40.3159Z" />
+            <path
+                d="M26.1332 24.5956L26.1332 24.5956L26.1445 24.5998L32.7407 27.026C32.7409 27.026 32.7411 27.0261 32.7412 27.0261C32.8858 27.0796 33.0105 27.176 33.0986 27.3024C33.1869 27.4289 33.2341 27.5794 33.2341 27.7336C33.2341 27.8879 33.1869 28.0384 33.0986 28.1649C33.0106 28.2913 32.8859 28.3877 32.7413 28.4411C32.7411 28.4412 32.7409 28.4413 32.7407 28.4413L26.1433 30.8675L26.1334 30.8711L26.1236 30.875C25.9576 30.9399 25.8069 31.0389 25.6814 31.1655C25.5559 31.2921 25.4582 31.4436 25.3946 31.6103L25.3946 31.6102L25.3904 31.6215L22.9639 38.2169C22.9639 38.2171 22.9638 38.2172 22.9638 38.2173C22.9103 38.3619 22.8139 38.4866 22.6875 38.5747C22.5609 38.6629 22.4103 38.7102 22.256 38.7102C22.1017 38.7102 21.951 38.6629 21.8245 38.5747C21.698 38.4866 21.6017 38.3619 21.5482 38.2175C21.5482 38.2173 21.5481 38.2171 21.548 38.2169L19.1216 31.6203L19.1179 31.6104L19.1141 31.6006C19.0491 31.4346 18.9501 31.2839 18.8235 31.1584C18.6968 31.0329 18.5453 30.9352 18.3787 30.8717L18.3787 30.8717L18.3675 30.8675L11.7712 28.4413C11.6264 28.3879 11.5015 28.2915 11.4133 28.1649C11.3251 28.0384 11.2778 27.8879 11.2778 27.7336C11.2778 27.5794 11.3251 27.4289 11.4133 27.3024C11.5014 27.1759 11.6262 27.0795 11.7709 27.0261C11.771 27.026 11.7711 27.026 11.7712 27.026L18.3687 24.5998L18.3785 24.5962L18.3883 24.5923C18.5544 24.5274 18.705 24.4284 18.8306 24.3018C18.9561 24.1751 19.0538 24.0236 19.1173 23.857L19.1174 23.8571L19.1215 23.8458L21.548 17.2504C21.5481 17.2503 21.5481 17.2501 21.5482 17.25C21.6016 17.1055 21.698 16.9807 21.8245 16.8926C21.951 16.8044 22.1016 16.7571 22.256 16.7571C22.4103 16.7571 22.5609 16.8044 22.6875 16.8926C22.814 16.9807 22.9104 17.1055 22.9638 17.2501C22.9638 17.2502 22.9639 17.2503 22.9639 17.2504L25.3904 23.847L25.394 23.8569L25.3979 23.8666C25.4628 24.0327 25.5618 24.1833 25.6885 24.3089C25.8151 24.4344 25.9666 24.532 26.1332 24.5956Z"
+                stroke="#010925"
+                stroke-width="2.04683"
+            />
+            <mask id="path-3-inside-1_55_242" fill="white">
+                <path d="M26.2044 14.2502C26.0544 14.4003 25.9701 14.6037 25.9701 14.8159C25.9701 15.0281 26.0544 15.2316 26.2044 15.3816C26.3545 15.5316 26.558 15.6159 26.7702 15.6159H29.1705V18.0159C29.1705 18.2281 29.2548 18.4316 29.4049 18.5816C29.5549 18.7316 29.7584 18.8159 29.9706 18.8159C30.1828 18.8159 30.3863 18.7316 30.5364 18.5816C30.6864 18.4316 30.7707 18.2281 30.7707 18.0159V15.6159H33.171C33.3832 15.6159 33.5867 15.5316 33.7368 15.3816C33.8868 15.2316 33.9711 15.0281 33.9711 14.8159C33.9711 14.6037 33.8868 14.4003 33.7368 14.2502C33.5867 14.1002 33.3832 14.0159 33.171 14.0159H30.7707V11.6159C30.7707 11.4037 30.6864 11.2003 30.5364 11.0502C30.3863 10.9002 30.1828 10.8159 29.9706 10.8159C29.7584 10.8159 29.5549 10.9002 29.4049 11.0502C29.2548 11.2003 29.1705 11.4037 29.1705 11.6159V14.0159H26.7702C26.558 14.0159 26.3545 14.1002 26.2044 14.2502Z" />
+            </mask>
+            <path
+                d="M26.2044 14.2502L27.6517 15.6977L27.6517 15.6977L26.2044 14.2502ZM29.1705 14.0159V16.0627H31.2173V14.0159H29.1705ZM29.4049 11.0502L27.9576 9.60281L27.9576 9.60281L29.4049 11.0502ZM30.5364 11.0502L29.0891 12.4977L29.0891 12.4977L30.5364 11.0502ZM30.7707 14.0159H28.7239V16.0627H30.7707V14.0159ZM33.7368 14.2502L35.184 12.8028L35.184 12.8028L33.7368 14.2502ZM33.7368 15.3816L35.184 16.829L35.184 16.829L33.7368 15.3816ZM30.7707 15.6159V13.5691H28.7239V15.6159H30.7707ZM29.4049 18.5816L27.9576 20.029L27.9576 20.029L29.4049 18.5816ZM29.1705 15.6159H31.2173V13.5691H29.1705V15.6159ZM28.0169 14.8159C28.0169 15.1467 27.8855 15.4638 27.6517 15.6977L24.7572 12.8028C24.2233 13.3367 23.9233 14.0608 23.9233 14.8159H28.0169ZM27.6517 15.6977C27.4178 15.9314 27.1008 16.0627 26.7702 16.0627V11.9691C26.0152 11.9691 25.2911 12.269 24.7572 12.8028L27.6517 15.6977ZM26.7702 16.0627H29.1705V11.9691H26.7702V16.0627ZM31.2173 14.0159V11.6159H27.1237V14.0159H31.2173ZM31.2173 11.6159C31.2173 11.9467 31.0859 12.2638 30.8521 12.4977L27.9576 9.60281C27.4237 10.1367 27.1237 10.8608 27.1237 11.6159H31.2173ZM30.8521 12.4977C30.6183 12.7315 30.3012 12.8627 29.9706 12.8627V8.76909C29.2157 8.76909 28.4916 9.06896 27.9576 9.60281L30.8521 12.4977ZM29.9706 12.8627C29.64 12.8627 29.323 12.7314 29.0891 12.4977L31.9836 9.60281C31.4497 9.06896 30.7256 8.76909 29.9706 8.76909V12.8627ZM29.0891 12.4977C28.8553 12.2638 28.7239 11.9467 28.7239 11.6159H32.8175C32.8175 10.8608 32.5175 10.1367 31.9836 9.60281L29.0891 12.4977ZM28.7239 11.6159V14.0159H32.8175V11.6159H28.7239ZM30.7707 16.0627H33.171V11.9691H30.7707V16.0627ZM33.171 16.0627C32.8405 16.0627 32.5234 15.9314 32.2896 15.6977L35.184 12.8028C34.6501 12.269 33.926 11.9691 33.171 11.9691V16.0627ZM32.2896 15.6977C32.0557 15.4638 31.9243 15.1467 31.9243 14.8159H36.018C36.018 14.0608 35.718 13.3367 35.184 12.8028L32.2896 15.6977ZM31.9243 14.8159C31.9243 14.4852 32.0557 14.168 32.2896 13.9342L35.184 16.829C35.718 16.2952 36.018 15.571 36.018 14.8159H31.9243ZM32.2896 13.9342C32.5234 13.7004 32.8405 13.5691 33.171 13.5691V17.6627C33.926 17.6627 34.6501 17.3629 35.184 16.829L32.2896 13.9342ZM33.171 13.5691H30.7707V17.6627H33.171V13.5691ZM28.7239 15.6159V18.0159H32.8175V15.6159H28.7239ZM28.7239 18.0159C28.7239 17.6852 28.8553 17.368 29.0891 17.1342L31.9836 20.029C32.5175 19.4952 32.8175 18.771 32.8175 18.0159H28.7239ZM29.0891 17.1342C29.323 16.9004 29.64 16.7691 29.9706 16.7691V20.8627C30.7256 20.8627 31.4497 20.5629 31.9836 20.029L29.0891 17.1342ZM29.9706 16.7691C30.3012 16.7691 30.6183 16.9004 30.8521 17.1342L27.9576 20.029C28.4916 20.5629 29.2157 20.8627 29.9706 20.8627V16.7691ZM30.8521 17.1342C31.0859 17.368 31.2173 17.6852 31.2173 18.0159H27.1237C27.1237 18.771 27.4237 19.4952 27.9576 20.029L30.8521 17.1342ZM31.2173 18.0159V15.6159H27.1237V18.0159H31.2173ZM29.1705 13.5691H26.7702V17.6627H29.1705V13.5691ZM26.7702 13.5691C27.1008 13.5691 27.4178 13.7004 27.6517 13.9342L24.7572 16.829C25.2911 17.3629 26.0152 17.6627 26.7702 17.6627V13.5691ZM27.6517 13.9342C27.8855 14.168 28.0169 14.4852 28.0169 14.8159H23.9233C23.9233 15.571 24.2233 16.2952 24.7572 16.829L27.6517 13.9342Z"
+                mask="url(#path-3-inside-1_55_242)"
+            />
+            <mask id="path-5-inside-2_55_242" fill="white">
+                <path d="M32.3716 20.5331C32.2965 20.6081 32.2544 20.7098 32.2544 20.8159C32.2544 20.922 32.2965 21.0237 32.3716 21.0988C32.4466 21.1738 32.5483 21.2159 32.6544 21.2159H33.8546V22.4159C33.8546 22.522 33.8968 22.6237 33.9718 22.6988C34.0468 22.7738 34.1486 22.8159 34.2547 22.8159C34.3608 22.8159 34.4625 22.7738 34.5375 22.6988C34.6126 22.6237 34.6547 22.522 34.6547 22.4159V21.2159H35.8549C35.961 21.2159 36.0627 21.1738 36.1377 21.0988C36.2128 21.0237 36.2549 20.922 36.2549 20.8159C36.2549 20.7098 36.2128 20.6081 36.1377 20.5331C36.0627 20.4581 35.961 20.4159 35.8549 20.4159H34.6547V19.2159C34.6547 19.1098 34.6126 19.0081 34.5375 18.9331C34.4625 18.8581 34.3608 18.8159 34.2547 18.8159C34.1486 18.8159 34.0468 18.8581 33.9718 18.9331C33.8968 19.0081 33.8546 19.1098 33.8546 19.2159V20.4159H32.6544C32.5483 20.4159 32.4466 20.4581 32.3716 20.5331Z" />
+            </mask>
+            <path
+                d="M32.3716 20.5331L33.8188 21.9805L33.8188 21.9805L32.3716 20.5331ZM33.8546 20.4159V22.4627H35.9014V20.4159H33.8546ZM33.9718 18.9331L32.5245 17.4857L32.5245 17.4857L33.9718 18.9331ZM34.5375 18.9331L33.0903 20.3805L33.0903 20.3805L34.5375 18.9331ZM34.6547 20.4159H32.6079V22.4627H34.6547V20.4159ZM36.1377 20.5331L37.585 19.0857L37.585 19.0857L36.1377 20.5331ZM36.1377 21.0988L37.585 22.5462L37.585 22.5462L36.1377 21.0988ZM34.6547 21.2159V19.1691H32.6079V21.2159H34.6547ZM33.9718 22.6988L32.5245 24.1462L32.5245 24.1462L33.9718 22.6988ZM33.8546 21.2159H35.9014V19.1691H33.8546V21.2159ZM34.3012 20.8159C34.3012 21.2528 34.1277 21.6717 33.8188 21.9805L30.9243 19.0857C30.4654 19.5445 30.2076 20.1669 30.2076 20.8159H34.3012ZM33.8188 21.9805C33.5099 22.2893 33.0911 22.4627 32.6544 22.4627V18.3691C32.0056 18.3691 31.3832 18.6268 30.9243 19.0857L33.8188 21.9805ZM32.6544 22.4627H33.8546V18.3691H32.6544V22.4627ZM35.9014 20.4159V19.2159H31.8078V20.4159H35.9014ZM35.9014 19.2159C35.9014 19.6528 35.7279 20.0717 35.419 20.3805L32.5245 17.4857C32.0656 17.9445 31.8078 18.5669 31.8078 19.2159H35.9014ZM35.419 20.3805C35.1102 20.6893 34.6913 20.8627 34.2547 20.8627V16.7691C33.6058 16.7691 32.9835 17.0268 32.5245 17.4857L35.419 20.3805ZM34.2547 20.8627C33.818 20.8627 33.3992 20.6893 33.0903 20.3805L35.9848 17.4857C35.5259 17.0268 34.9035 16.7691 34.2547 16.7691V20.8627ZM33.0903 20.3805C32.7814 20.0717 32.6079 19.6528 32.6079 19.2159H36.7015C36.7015 18.5669 36.4437 17.9445 35.9848 17.4857L33.0903 20.3805ZM32.6079 19.2159V20.4159H36.7015V19.2159H32.6079ZM34.6547 22.4627H35.8549V18.3691H34.6547V22.4627ZM35.8549 22.4627C35.4182 22.4627 34.9994 22.2893 34.6905 21.9805L37.585 19.0857C37.1261 18.6268 36.5037 18.3691 35.8549 18.3691V22.4627ZM34.6905 21.9805C34.3817 21.6717 34.2081 21.2528 34.2081 20.8159H38.3017C38.3017 20.1669 38.0439 19.5445 37.585 19.0857L34.6905 21.9805ZM34.2081 20.8159C34.2081 20.3791 34.3817 19.9602 34.6905 19.6513L37.585 22.5462C38.0439 22.0873 38.3017 21.4649 38.3017 20.8159H34.2081ZM34.6905 19.6513C34.9994 19.3425 35.4182 19.1691 35.8549 19.1691V23.2627C36.5037 23.2627 37.1261 23.005 37.585 22.5462L34.6905 19.6513ZM35.8549 19.1691H34.6547V23.2627H35.8549V19.1691ZM32.6079 21.2159V22.4159H36.7015V21.2159H32.6079ZM32.6079 22.4159C32.6079 21.9791 32.7814 21.5602 33.0903 21.2513L35.9848 24.1462C36.4437 23.6873 36.7015 23.0649 36.7015 22.4159H32.6079ZM33.0903 21.2513C33.3992 20.9425 33.818 20.7691 34.2547 20.7691V24.8627C34.9035 24.8627 35.5259 24.605 35.9848 24.1462L33.0903 21.2513ZM34.2547 20.7691C34.6913 20.7691 35.1102 20.9425 35.419 21.2513L32.5245 24.1462C32.9835 24.605 33.6058 24.8627 34.2547 24.8627V20.7691ZM35.419 21.2513C35.7279 21.5602 35.9014 21.9791 35.9014 22.4159H31.8078C31.8078 23.0649 32.0656 23.6873 32.5245 24.1462L35.419 21.2513ZM35.9014 22.4159V21.2159H31.8078V22.4159H35.9014ZM33.8546 19.1691H32.6544V23.2627H33.8546V19.1691ZM32.6544 19.1691C33.0911 19.1691 33.5099 19.3425 33.8188 19.6513L30.9243 22.5462C31.3832 23.005 32.0056 23.2627 32.6544 23.2627V19.1691ZM33.8188 19.6513C34.1277 19.9602 34.3012 20.3791 34.3012 20.8159H30.2076C30.2076 21.4649 30.4654 22.0873 30.9243 22.5462L33.8188 19.6513Z"
+                mask="url(#path-5-inside-2_55_242)"
+            />
+        </svg>
+    );
 }
